+++ lib/app/error/error_model.dart
