+++ lib/app/error/result.dart
