+++ lib/app/router/auth_route_interceptor.dart
@@ -1,10 +1,11 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter_best_practice/app/router/routes.dart';
-import 'package:flutter_best_practice/core/route/app_router_interceptor.dart';
 import 'package:flutter_best_practice/ui/controller/auth/auth_controller.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
+
+import '../foundation/route/app_router_interceptor.dart';
 
 part 'generated/auth_route_interceptor.g.dart';
