+++ lib/domain/usecase/auth/sign_in_use_case.dart
@@ -1,9 +1,8 @@
 import 'package:flutter_best_practice/domain/repository/auth_repository.dart';
 import 'package:flutter_best_practice/domain/model/auth/sign_in_model.dart';
-
-import '../../../core/error_handling/custom_exception.dart';
-import '../../../core/error_handling/result.dart';
-import '../../../core/usecase/use_case.dart';
+import '../../../app/error/custom_exception.dart';
+import '../../../app/error/result.dart';
+import '../../../app/foundation/usecase/use_case.dart';
 import '../../param/auth/sign_in_params.dart';
