+++ lib/domain/usecase/sample/get_samples_use_case.dart
@@ -1,9 +1,9 @@
-import 'package:flutter_best_practice/core/error_handling/custom_exception.dart';
-import 'package:flutter_best_practice/core/error_handling/result.dart';
-import 'package:flutter_best_practice/core/usecase/no_params.dart';
-import 'package:flutter_best_practice/core/usecase/use_case.dart';
 import 'package:flutter_best_practice/domain/model/sample/sample_model.dart';
 import 'package:flutter_best_practice/domain/repository/sample_repository.dart';
+import '../../../app/error/custom_exception.dart';
+import '../../../app/error/result.dart';
+import '../../../app/foundation/usecase/no_params.dart';
+import '../../../app/foundation/usecase/use_case.dart';
 
 class GetSamplesUseCase implements UseCase<List<SampleModel>, NoParams> {
   final SampleRepository repository;
