+++ lib/domain/usecase/task/get_tasks_use_case.dart
@@ -1,8 +1,9 @@
-import 'package:flutter_best_practice/core/error_handling/result.dart';
-import 'package:flutter_best_practice/core/usecase/no_params.dart';
-import 'package:flutter_best_practice/core/usecase/use_case.dart';
 import 'package:flutter_best_practice/domain/model/task/task_model.dart';
 import 'package:flutter_best_practice/domain/repository/task_repository.dart';
+
+import '../../../app/error/result.dart';
+import '../../../app/foundation/usecase/no_params.dart';
+import '../../../app/foundation/usecase/use_case.dart';
 
 class GetTasksUseCase implements UseCase<List<TaskModel>, NoParams> {
   final TaskRepository _repository;
