+++ lib/ui/view/auth/sign_in/sign_in_state.dart
@@ -1,7 +1,7 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-import 'package:flutter_best_practice/core/error_handling/custom_exception.dart';
-import 'package:flutter_best_practice/core/status/status.dart';
+import '../../../../app/error/custom_exception.dart';
+import '../../../../app/foundation/status/status.dart';
 part 'generated/sign_in_state.freezed.dart';
