+++ lib/ui/view/auth/sign_in/sign_in_view.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_best_practice/core/status/status.dart';
+import 'package:flutter_best_practice/app/foundation/status/status.dart';
 import 'package:flutter_best_practice/ui/view/auth/sign_in/sign_in_controller.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
