+++ lib/ui/view/sample/sample_state.dart
@@ -1,9 +1,8 @@
 import 'package:flutter_best_practice/domain/model/sample/sample_model.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-import 'package:flutter_best_practice/core/error_handling/custom_exception.dart';
-import 'package:flutter_best_practice/core/status/status.dart';
-
+import '../../../app/error/custom_exception.dart';
+import '../../../app/foundation/status/status.dart';
 part 'generated/sample_state.freezed.dart';
 
 /// sample view의 상태를 관리하기 위한 state class
