+++ lib/ui/view/task/add/add_task_state.dart
@@ -1,8 +1,7 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-import '../../../../core/error_handling/custom_exception.dart';
-import '../../../../core/status/status.dart';
-
+import '../../../../app/error/custom_exception.dart';
+import '../../../../app/foundation/status/status.dart';
 part 'generated/add_task_state.freezed.dart';
 
 /// add task view의 상태를 관리하기 위한 state class
