+++ lib/ui/view/task/task_view.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_best_practice/app/foundation/status/status.dart';
 import 'package:flutter_best_practice/app/router/routes.dart';
-import 'package:flutter_best_practice/core/status/status.dart';
 import 'package:flutter_best_practice/ui/view/task/task_controller.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
